+++ src/fetch-client-ui/components/ResponseUI/ResponsePanel/consts.ts
@@ -2,6 +2,7 @@ export const responseOptions = [
 	{ name: "Response", value: "response", },
 	{ name: "Headers", value: "headers", },
 	{ name: "Cookies", value: "cookies", },
+	{ name: "PreFetch Results", value: "prefetchresults", },
 	{ name: "Test Results", value: "testresults", },
 	{ name: "Notes", value: "notes", }
 ];