+++ src/utils/PreFetchRunner.ts
@@ -1,6 +1,6 @@
 import { IPreFetch, IRequestModel } from "../fetch-client-ui/components/RequestUI/redux/types";
 import { InitialResponse } from "../fetch-client-ui/components/ResponseUI/redux/reducer";
-import { IReponseModel } from "../fetch-client-ui/components/ResponseUI/redux/types";
+import { IPreFetchResponse, IReponseModel } from "../fetch-client-ui/components/ResponseUI/redux/types";
 import { ISettings, IVariable } from "../fetch-client-ui/components/SideBar/redux/types";
 import { executeTests, setVariable } from "../fetch-client-ui/components/TestUI/TestPanel/helper";
 import { GetParentSettingsSync, GetVariableByColId } from "./db/collectionDBUtil";
@@ -12,7 +12,7 @@ import { apiFetch, FetchConfig } from "./fetchUtil";
 export class PreFetchRunner {
 	private readonly fetchConfig: FetchConfig;
 	private executingRequests: string[] = [];
-	private allow: boolean = true;
+	public allow: boolean = true;
 	private response: IReponseModel = {
 		response: JSON.parse(JSON.stringify(InitialResponse)),
 		headers: [],
@@ -20,6 +20,7 @@ export class PreFetchRunner {
 		loading: false,
 		testResults: []
 	};
+	public preFetchResponses: IPreFetchResponse[] = [];
 	public message: string = "";
 
 
@@ -28,7 +29,7 @@ export class PreFetchRunner {
 		this.fetchConfig = fetchConfig;
 	}
 
-	RunPreRequests = async (preFetch: IPreFetch, fetchIndex: number, reqIndex: number, parentName: string, isCollectionPreRequest: boolean) => {
+	RunPreRequests = async (preFetch: IPreFetch, reqIndex: number, parentName: string, isCollectionPreRequest: boolean, parentIndex: number = -1, parentPreFetchResponse: IPreFetchResponse = null) => {
 		let request: IRequestModel;
 		let parentSettings: ISettings;
 		let updateVariable: IVariable;
@@ -47,16 +48,51 @@ export class PreFetchRunner {
 			this.executingRequests.push(preFetch.requests[reqIndex].reqId);
 		}
 
-		for (let i = 0; i < preFetch.requests.length && this.allow; i++) {
+		let filterPreFetchRequests = preFetch.requests?.filter(i => i.reqId && i.reqId !== "undefined");
+
+		for (let i = 0; i < filterPreFetchRequests.length && this.allow; i++) {
+
+			if (parentIndex === -1) {
+				this.preFetchResponses.push({
+					reqId: "",
+					name: "",
+					resStatus: 0,
+					testResults: [],
+					childrenResponse: []
+				});
+			} else {
+				parentPreFetchResponse.childrenResponse.push({
+					reqId: "",
+					name: "",
+					resStatus: 0,
+					testResults: [],
+					childrenResponse: []
+				});
+			}
+
 			reqIndex = i;
+
 			if (i > 0 && request?.id) {
-				let condition = preFetch.requests[i].condition;
-				for (let j = 0; j < condition.length; j++) {
-					if (!condition[j].parameter || !condition[j].action) {
-						break;
-					}
+				let condition = filterPreFetchRequests[i].condition.filter(i => i.parameter && i.action);
+				if (condition?.length > 0) {
 					let testResult = executeTests(condition, this.response, updateVariable?.data);
-					if (testResult.findIndex(i => i.result === false) !== -1) {
+					let failedResult = testResult.findIndex(it => it.result === false);
+
+					if (parentIndex === -1) {
+						this.preFetchResponses[i].testResults = testResult;
+						if (failedResult !== -1) {
+							this.preFetchResponses[i].reqId = "-1";
+						}
+					} else {
+						let childIndex = parentPreFetchResponse.childrenResponse?.length;
+						parentPreFetchResponse.childrenResponse[childIndex - 1].testResults = testResult;
+						if (failedResult !== -1) {
+							parentPreFetchResponse.reqId = "-1";
+							parentPreFetchResponse.childrenResponse[childIndex - 1].reqId = "-1";
+						}
+					}
+
+					if (failedResult !== -1) {
 						this.allow = false;
 						this.message = `'Condition ${reqIndex}' failed in 'Pre-Request ${reqIndex}' in the Request '${parentName}'`;
 						return;
@@ -64,9 +100,9 @@ export class PreFetchRunner {
 				}
 			}
 
-			let reqId = preFetch.requests[i].reqId;
-			let parentId = preFetch.requests[i].parentId;
-			let colId = preFetch.requests[i].colId;
+			let reqId = filterPreFetchRequests[i].reqId;
+			let parentId = filterPreFetchRequests[i].parentId;
+			let colId = filterPreFetchRequests[i].colId;
 
 			if (!reqId) {
 				return;
@@ -84,25 +120,73 @@ export class PreFetchRunner {
 			request = await GetRequestItem(reqId);
 
 			if (request && this.allow) {
+
+				if (parentIndex !== -1) {
+					parentPreFetchResponse.childrenResponse[i].name = request.name;
+				} else {
+					this.preFetchResponses[i].name = request.name;
+				}
+
 				if (request.preFetch?.requests?.length > 0 && !isCollectionPreRequest) {
-					await this.RunPreRequests(request.preFetch, fetchIndex + 1, 0, request.name, isCollectionPreRequest);
+
+					await this.RunPreRequests(request.preFetch, 0, request.name, isCollectionPreRequest, reqIndex, parentIndex === -1 ? this.preFetchResponses[i] : parentPreFetchResponse.childrenResponse[i]);
 					if (!this.allow) {
 						return;
 					}
+
 					let index = this.executingRequests.findIndex(i => i === reqId);
 					if (index !== -1) {
 						this.executingRequests.splice(index);
 					}
+
 					let res = await apiFetch(request, variable?.data, parentSettings, null, this.fetchConfig);
+
+					if (parentIndex !== -1) {
+						parentPreFetchResponse.childrenResponse[i].reqId = request.id;
+						parentPreFetchResponse.childrenResponse[i].resStatus = res?.response?.status;
+					} else {
+						this.preFetchResponses[i].reqId = request.id;
+						this.preFetchResponses[i].resStatus = res?.response?.status;
+					}
+
+					if (res && res.response?.status >= 400 && res.response?.status <= 599) {
+						if (parentIndex !== -1) {
+							parentPreFetchResponse.reqId = "-1";
+						}
+						this.allow = false;
+						this.message = `Pre-Request '${request.name}' failed in the Request '${parentName}'`;
+						return;
+					}
 					this.response.response = res.response;
 					this.response.headers = res.headers;
 					this.response.cookies = res.cookies;
+
 					updateVariable = await this.updateVariable(request, variable);
 				} else {
 					let res = await apiFetch(request, variable?.data, parentSettings, null, this.fetchConfig);
+
+					if (parentIndex !== -1) {
+						parentPreFetchResponse.childrenResponse[i].reqId = request.id;
+						parentPreFetchResponse.childrenResponse[i].name = request.name;
+						parentPreFetchResponse.childrenResponse[i].resStatus = res?.response?.status;
+					} else {
+						this.preFetchResponses[i].reqId = request.id;
+						this.preFetchResponses[i].name = request.name;
+						this.preFetchResponses[i].resStatus = res?.response?.status;
+					}
+
+					if (res && res.response?.status >= 400 && res.response?.status <= 599) {
+						if (parentIndex !== -1) {
+							parentPreFetchResponse.reqId = "-1";
+						}
+						this.allow = false;
+						this.message = `Pre-Request '${request.name}' failed in the Request '${parentName}'`;
+						return;
+					}
 					this.response.response = res.response;
 					this.response.headers = res.headers;
 					this.response.cookies = res.cookies;
+
 					updateVariable = await this.updateVariable(request, variable);
 				}
 			}